EXP: thread: provide an alternative to pthread's rwlock
Since version 1.1.0, OpenSSL's libcrypto ignores the provided locking
mechanism and uses pthread's rwlocks instead. The problem is that this
results in a huge amount of syscalls on systems facing a bit of contention,
to the point where more than 80% of the CPU can be spent in the system
dealing with spinlocks just for futex_wake().

This patch provides an alternative by redefining the relevant pthread
rwlocks from the progressive locks. This way there will be no more
syscalls in case of contention, and CPU will be burnt in userland.

The pthread_rwlock API is complete except that the timed versions of
the rdlock and wrlock do not wait and simply fall back to trylock
versions.

The option can be enabled at build time using USE_PTHREAD_EMULATION=1.